Этап 4 - Система поиска. Реализовал систему поиска информации с использованием созданного поискового индекса. - реализовал метод '/search' (для одного сайта) - рефаторинг очистки таблиц, заменил на TRUNCATE - рефакторинг метода очистки таблиц для метода '/indexPage' - рефакторинг метода SinglePageProcessor.processSinglePage - исправил ошибки подсчёта лемм и синхронизации потоков

+++ src/main/java/searchengine/dao/model/Index.java
@@ -34,7 +34,7 @@ public class Index {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name = "page_id", nullable = false)
@@ -45,5 +45,5 @@ public class Index {
     private Lemma lemma;
 
     @Column(name = "`rank`", columnDefinition = "FLOAT", nullable = false)
-    private Float rank;
+    private Integer rank;
 }
